### 2025-08-04 f914ba4
Update paper.md
