### 2025-08-08 #8744 @Ahmedbargheet
Dear @crvernon   just following up to see if there’s anything else needed from my side. I noticed that the editorialbot didn’t respond after the last check, and no editor has been assigned yet. Please let me know if I should take any further action. Thanks again!
