### 2025-08-08 3f5f47b
Update paper.md
