### 2025-08-08 2b9a1a6
Update paper.md
