### 2025-08-08 301211c
paper.md
