### 2025-08-08 887c398
paper.md
