### 2025-08-08 4c2bc9b
paper.md
