### 2025-08-08 7e31529
paper.md
