+++ paper/paper.bib
@@ -64,7 +64,6 @@
   title={Testing for normality: a user’s (cautionary) guide},
   author={Gosselin, Romain-Daniel},
   journal={Laboratory animals},
-  doi={10.1177/002367722412768},
   url={https://journals.sagepub.com/doi/full/10.1177/00236772241276808},
   volume={58},
   number={5},
