### 2025-08-08 fd8734e
paper.md
