+++ paper/paper.md
@@ -51,8 +51,8 @@ A researcher uploads a CSV dataset, imported using the readr package [@wickham20
 ![](figure2.png)
 Figure 2: The assumption checking report.
 
-<p align="center">
-  <img src="figure3.png" width="90%">
+<p align="left">
+  <img src="figure3.png" width="100%">
 </p>
 
 <p align="left"><b>Figure 3:</b> Kruskal-Wallis test and effect sizes estimation using Kruskal-Wallis eta-squared.</p>
