### 2025-08-08 659d3ba
paper.md
