+++ paper/paper.md
@@ -29,7 +29,7 @@ link-citations: yes
 # Statement of Need
 Rigorous statistical analysis depends on checking underlying assumptions and choosing appropriate tests. Yet in practice, these steps are often skipped, misunderstood, or incorrectly applied. Reviews of social science literature report that fewer than 25% of studies explicitly mention assumption checks, and even fewer report them correctly [@shatz2024assumption]. This contributes to widespread misuse of statistical tests and undermines reproducibility. One common issue is the misinterpretation of assumption tests. For example, many users mistakenly interpret a significant Shapiro–Wilk test result (*p* < 0.05) as evidence for normality rather than deviation from it [@ghasemi2012normality; @midway2025testing; @gosselin2024testing]. Moreover, in small samples, where the Shapiro–Wilk test lacks power, borderline results are easily overlooked or misjudged.
 
-`AssumpSure` directly addresses these problems by making assumption checking a required part of the analysis workflow. Through an intuitive web interface, users are guided step-by-step: assumptions are tested automatically, results are color-coded (green for met, red for violated, orange for borderline), and clear recommendations are given. For example, if ANOVA assumptions are not met, Kruskal–Wallis is suggested. By complementing color-coded squares with tooltips and clear messages, `AssumpSure` helps reduce misinterpretation of common assumption tests and guide the user how to proceed.
+`AssumpSure` directly addresses these problems by making assumption checking a required part of the analysis workflow. Through an intuitive web interface, users are guided step-by-step: assumptions are tested automatically, results are color-coded (green for met, red for violated, orange for borderline), and clear recommendations are given. For example, if ANOVA assumptions are not met, Kruskal–Wallis is suggested. By complementing color-coded squares with tooltips and clear messages, `AssumpSure` helps reduce misinterpretation of common assumption tests and guides the user on how to proceed.
 
 # Comparison to other Packages
 Tools like *Jamovi* [@csahin2019jamovi] and *JASP* [@love2019jasp] offer user-friendly interfaces but lack integrated workflows that enforce mandatory assumption diagnostics, appropriate test selection, and actionable recommendations, leaving users to manually interpret complex outputs. This increases errors and undermines reproducibility. Additionally, non-R applications such as assumption-sheriff (Python) [@assumptionsheriff] also aim to simplify statistical assumption checking and test selection. However, these alternatives lack fully integrated, interactive diagnostic workflows or do not provide enforced guidance on test appropriateness for each step.
@@ -37,26 +37,27 @@ Tools like *Jamovi* [@csahin2019jamovi] and *JASP* [@love2019jasp] offer user-fr
 # Software Description
 `AssumpSure`’s modular R Shiny architecture organizes statistical analyses into continuous variable tests, categorical tests, correlations, and regression models. Users upload CSV data, and the interface automatically detects variable types, preventing invalid selections with clear error messages (e.g., suggesting one-way ANOVA if a t-test is chosen for a categorical variable with three levels). Mandatory assumption diagnostics (normality via Shapiro-Wilk, variance homogeneity via Levene’s test) are visualized through diagnostic plots (QQ plots, histograms, scatter plots), with tooltips to guide the user how to interpret the plots.
 
[…]
