### 2025-08-08 #8744 @Ahmedbargheet
> **Submitting author:** [@Ahmedbargheet](https://github.com/Ahmedbargheet) ([Ahmed Bargheet](http://orcid.org/0000-0002-0363-0249)) **Repository:** https://github.com/Ahmedbargheet/AssumpSure **Branch with paper.md** (empty if default branch): paper **Version:** v.1.1.0 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Chris Vernon
>
> ## Status
> [![status](https://camo.githubusercontent.com/37b9489b42b256d14d4b27ec1186fb6ac1d641b36248ce70bf3d73c34e36a307/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f63313634316661643330643562306439353137373130333138326532313661392f7374617475732e737667)](https://joss.theoj.org/papers/c1641fad30d5b0d95177103182e216a9)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS [@Ahmedbargheet](https://github.com/Ahmedbargheet). **Currently, there isn't a JOSS editor assigned** to your paper.
>
> [@Ahmedbargheet](https://github.com/Ahmedbargheet) if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot [@editorialbot](https://github.com/editorialbot) is here to help you find and assign reviewers and start the main review. To find out what [@editorialbot](https://github.com/editorialbot) can do for you type:
>
> ```
> @editorialbot commands
> ```

According to the comment above, I can suggest 4 potential reviewers:

Stephanie Pennington
Cedric Midoux
Biwei Cao
Salix Dubois
