+++ cran-comments.md
@@ -1,26 +1,0 @@
-Hello CRAN Team,
-
-This is an update for AssumpSure package (version 1.1.0) following the first release (Version 1.0.0).
-
-**Why I am submitting this update:**
-
-After release, users and I found several important bugs that could mislead users or result in incorrect statistical advice:
-
-- In some cases, the app suggested the wrong statistical test.
-- For small sample sizes, the app gave inappropriate guidance about normality testing.
-- Allowed Chi-square test for categorical variables with only one level, which is not statistically valid.
-- When a covariate showed perfect separation in the data, the app could not plot results, and users were not properly informed of the modeling limitations.
-
-These are serious issues, but they are now fixed and users are properly notified.
-
-
-**Additional improvements:**
-While addressing these bugs, I also added new features and improved user guidance. See NEWS file for details.
-
-**Checks:**
-- All R CMD checks passed
-
-Thank you for your understanding.
-
-Best regards,  
-Ahmed Bargheet
