+++ inst/app/app.R
@@ -4179,6 +4179,13 @@ output$boxplot_ui <- renderUI({
       session$sendCustomMessage("reinit-tooltips", list())
     }, once = FALSE)
   })
+  
+  # Helper: did outlier screen pass?
+  cor_outliers_passed <- reactive({
+    res <- car_outliers_val()
+    is.list(res) && !is.null(res$status) && identical(res$status, "clean")
+  })
+  
 
 
 # Data Loading & Numeric Feature Detection
@@ -4496,12 +4503,51 @@ output$boxplot_ui <- renderUI({
     mvn_ok_val(mvn_ok)
 
 
-    # --- Outliers (car::outlierTest)
+    # --- Outliers (Mahalanobis bivariate screen for Pearson)
     lm_model <- lm(v2 ~ v1)
-    car_outliers <- tryCatch({
-      car::outlierTest(lm_model)
-    }, error = function(e) NULL)
-    car_outliers_val(car_outliers)
+    alpha_out <- 0.01
+    df2_maha <- df %>%
+      dplyr::select(all_of(input$cor_features)) %>%
+      stats::na.omit() %>%
+      as.data.frame()
+    
+    if (nrow(df2_maha) < 5) {
+      car_outliers_val(list(
+        status = "na",
+        alpha = alpha_out,
+        n = NA_integer_,
+        idx = integer(0),
+        D2 = numeric(0),
+        p = numeric(0),
+        message = "Screen not applicable: fewer than 5 complete observations."
+      ))
+    } else {
+      S <- stats::cov(df2_maha)
+      if (!is.finite(det(S)) || abs(det(S)) < .Machine$double.eps) {
+        car_outliers_val(list(
+          status = "na",
+          alpha = alpha_out,
+          n = NA_integer_,
+          idx = integer(0),
+          D2 = numeric(0),
+          p = numeric(0),
+          message = "Screen not applicable: singular covariance."
+        ))
+      } else {
+        D2 <- stats::mahalanobis(df2_maha, colMeans(df2_maha), S)
+        pvals <- stats::pchisq(D2, df = 2, lower.tail = FALSE)
+        idx <- which(pvals < alpha_out)
+        car_outliers_val(list(
+          status = if (length(idx) > 0) "outliers" else "clean",
+          alpha = alpha_out,
+          n = length(idx),
+          idx = idx,
+          D2 = D2[idx],
+          p = pvals[idx],
+          message = NULL
+        ))
+      }
+    }
 
     # --- Homoscedasticity
     perf_het <- tryCatch({
@@ -4555,7 +4601,7 @@ output$boxplot_ui <- renderUI({
     class='fa-solid fa-circle-info'
     data-bs-toggle='tooltip'
     data-bs-placement='right'
-    title='This scatterplot shows the relationship between the two selected variables. The red line is a linear regression fit. If points follow the line, the relationship is linear.'
+    title='This scatterplot shows the relationship between the two selected variables. The red line represents a linear regression fit. If the points closely follow the line, the linearity assumption is reasonable; large deviations suggest the assumption may not hold.'
     style='color:#2c3e50; cursor:pointer; font-size:16px; margin-left:10px;'></i>"))
         ),
         fluidRow(
@@ -4617,43 +4663,19 @@ output$boxplot_ui <- renderUI({
 
 
         # 4. Outliers
-        h5(strong("4. Outlier Test")),
-        uiOutput("cor_outlier_text"),
+        # 4. Mahalanobis Outlier Test (Bivariate)
         div(
-          style = paste0(
-            "background-color:",
-            {
[…]
