+++ inst/app/app.R
@@ -1081,11 +1081,45 @@ server <- function(input, output, session) {
     req(df)
     
     # Continuous: numeric with many unique values
-    numeric_vars <- names(df)[sapply(df, function(col) {
-      is.numeric(col) &&
-        length(unique(col)) > 10 &&
-        any(col %% 1 != 0, na.rm = TRUE)  # Exclude if all values are integers
-    })]
+    # Continuous = numeric, not an ID, enough variety; be conservative with whole-number integers
+    numeric_vars <- names(df)[vapply(seq_along(df), function(i) {
+      col <- df[[i]]
+      nm  <- tolower(names(df)[i])
+      
+      if (!is.numeric(col)) return(FALSE)
+      x <- col[is.finite(col)]
+      n <- length(x); if (n < 3) return(FALSE)
+      u <- length(unique(x)); ur <- if (n > 0) u / n else 0
+      all_int <- all(abs(x - round(x)) < 1e-9)
+      
+      # 1) Drop obvious IDs by name or near-sequence
+      if (grepl("\\b(id|row|index)\\b", nm)) return(FALSE)
+      if (u == n && all_int && all(diff(sort(x)) == 1)) return(FALSE)  # strictly sequential unique integers
+      
+      # 1b) Drop integer ladders with low duplicates (e.g., 1,1,2,2,3,3,...)
+      if (all_int) {
+        uni <- sort(unique(x))
+        if (length(uni) >= 8 && all(diff(uni) == 1)) {
+          freq <- as.numeric(table(x))
+          if (median(freq) <= 2 && max(freq) <= 3) return(FALSE)
+        }
+      }
+      
+      # 2) Drop count-like integers with high uniqueness at small n
+      if (all_int && min(x, na.rm = TRUE) >= 0) {
+        freq <- table(x)
+        prop_singletons <- mean(freq == 1)   # share of values that appear once
+        if ((ur >= 0.8 || prop_singletons >= 0.7) && n <= 50) return(FALSE)
+      }
+      
+      # 3) Drop few-level integers → likely counts/ordinal
+      if (all_int && (u <= 5 || ur < 0.35)) return(FALSE)
+      
+      # 4) Keep if enough distinct values relative to n
+      return(u >= min(8, ceiling(0.5 * n)))
+    }, logical(1))]
+    
+    
     
     # Categorical: factor or character with few unique values
     cat_vars <- names(df)[sapply(df, function(col) {
@@ -1214,7 +1248,7 @@ server <- function(input, output, session) {
     }
   })
 
-
+  
   # reset Welch ANOVA
   observeEvent(input$file, { welch_anova_clicked(FALSE) })
   observeEvent(input$test_type, { welch_anova_clicked(FALSE) })
@@ -1363,6 +1397,9 @@ server <- function(input, output, session) {
   
   observeEvent(input$file, { welch_clicked(FALSE) })
   observeEvent(input$check_assumptions, { welch_clicked(FALSE) })
+  
+  
+  
   
 
  # Run the Post hoc test
@@ -1825,24 +1862,36 @@ output$levene_text <- renderUI({
 
   # One-way ANOVA
   qq_plot_anova <- function(df) {
-    ggplot2::ggplot(df, aes(sample = value, color = group)) +
-      stat_qq() +
-      stat_qq_line(color = "#E41A1C") +
-      theme_test() +
-      scale_color_brewer(palette = "Set2") +
-      facet_wrap(~group) +
-      theme(strip.text = element_text(size = 12, face = "bold", color = "black")) + 
-      theme(strip.background = element_rect(colour = "black", fill = "white")) + 
-      theme(axis.title.x = element_text(colour = "black", face="bold", size = 12)) +
-      theme(axis.title.y = element_text(colour = "black", face="bold", size = 12)) + 
[…]
