+++ inst/app/app.R
@@ -1067,6 +1067,10 @@ server <- function(input, output, session) {
         }
       })
       
+      # Warn if dataset is too small
+      if (nrow(df) < 10) {
+        showNotification(strong("Warning: The app only detects continuous variables with ≥10 observations and ≥5 distinct values. This is the minimum for detection; reliable statistics usually need ≥20–30 observations per group."), type = "warning", duration = 15)
+      }
       
       df
     }, error = function(e) {
@@ -1081,43 +1085,42 @@ server <- function(input, output, session) {
     req(df)
     
     # Continuous: numeric with many unique values
-    # Continuous = numeric, not an ID, enough variety; be conservative with whole-number integers
-    numeric_vars <- names(df)[vapply(seq_along(df), function(i) {
-      col <- df[[i]]
-      nm  <- tolower(names(df)[i])
+    # Final, production-ready replacement
+    # Returns names of continuous variables meeting data sufficiency and spread criteria.
+    get_continuous_vars <- function(df,
+                                    min_n = 10,          # minimum finite observations
+                                    min_unique = 5,     # minimum distinct finite values
+                                    exclude_pure_integers = TRUE,
+                                    fractional_tol = sqrt(.Machine$double.eps)) {
       
-      if (!is.numeric(col)) return(FALSE)
-      x <- col[is.finite(col)]
-      n <- length(x); if (n < 3) return(FALSE)
-      u <- length(unique(x)); ur <- if (n > 0) u / n else 0
-      all_int <- all(abs(x - round(x)) < 1e-9)
-      
-      # 1) Drop obvious IDs by name or near-sequence
-      if (grepl("\\b(id|row|index)\\b", nm)) return(FALSE)
-      if (u == n && all_int && all(diff(sort(x)) == 1)) return(FALSE)  # strictly sequential unique integers
-      
-      # 1b) Drop integer ladders with low duplicates (e.g., 1,1,2,2,3,3,...)
-      if (all_int) {
-        uni <- sort(unique(x))
-        if (length(uni) >= 8 && all(diff(uni) == 1)) {
-          freq <- as.numeric(table(x))
-          if (median(freq) <= 2 && max(freq) <= 3) return(FALSE)
-        }
+      # Fast unique counter
+      unique_count <- function(x) {
+        if (requireNamespace("data.table", quietly = TRUE)) data.table::uniqueN(x) else length(unique(x))
       }
       
-      # 2) Drop count-like integers with high uniqueness at small n
-      if (all_int && min(x, na.rm = TRUE) >= 0) {
-        freq <- table(x)
-        prop_singletons <- mean(freq == 1)   # share of values that appear once
-        if ((ur >= 0.8 || prop_singletons >= 0.7) && n <= 50) return(FALSE)
+      is_continuous <- function(col) {
+        if (!is.numeric(col)) return(FALSE)
+        
+        x <- col[is.finite(col)]
+        if (length(x) < min_n) return(FALSE)
+        
+        v <- stats::var(x)
+        if (is.na(v) || v <= 0) return(FALSE)
+        
+        if (unique_count(x) < min_unique) return(FALSE)
+        
+        if (exclude_pure_integers && all(abs(x - round(x)) <= fractional_tol)) return(FALSE)
+        
+        TRUE
       }
       
-      # 3) Drop few-level integers → likely counts/ordinal
[…]
