+++ DESCRIPTION
@@ -38,6 +38,9 @@ Imports:
     glmmTMB,
     effectsize,
     coin
+Depends: R (>= 4.2.0)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
+URL: https://github.com/Ahmedbargheet/AssumpSure
+BugReports: https://github.com/Ahmedbargheet/AssumpSure/issues
