### 2025-08-20 #8744 @Ahmedbargheet
Dear @crvernon, would it be possible to suggest a reviewer from outside the official JOSS reviewer list? I can provide their name and contact information for your evaluation, if that would help expedite the editor assignment process.
