+++ inst/app/app.R
@@ -7409,7 +7409,9 @@ lmm_vars <- reactive({
 
       log_data <- reactive({
         req(input$log_file)
-        readr::read_csv(input$log_file$datapath, show_col_types = F) %>%
+        
+        ## assign to df first
+        df <- readr::read_csv(input$log_file$datapath, show_col_types = FALSE) %>%
           dplyr::select(
             -dplyr::contains("infant_id", ignore.case = TRUE),
             -dplyr::contains("infantid", ignore.case = TRUE),
@@ -7419,6 +7421,23 @@ lmm_vars <- reactive({
             -dplyr::contains("accession", ignore.case = TRUE),
             -dplyr::contains("name", ignore.case = TRUE)
           )
+        
+        ## now safely transform df
+        df[] <- lapply(df, function(col) {
+          if (is.character(col)) {
+            factor(col)                          
+          } else if (is.logical(col)) {
+            factor(col)                          
+          } else if (is.numeric(col) &&
+                     all(col == floor(col), na.rm = TRUE) &&
+                     length(unique(col)) <= 15) {
+            factor(as.character(col))            
+          } else {
+            col
+          }
+        })
+        
+        df
       })
 
 
@@ -7496,6 +7515,8 @@ lmm_vars <- reactive({
 
         output$log_result <- renderUI({
           df <- log_model_data()
+          validate(need(!is.null(df), ""))
+          
           main_effects <- input$log_indep
           interaction_terms <- NULL
           if (!is.null(input$log_interact) && length(input$log_interact) > 0) {
@@ -7505,7 +7526,12 @@ lmm_vars <- reactive({
           fml <- as.formula(paste(input$log_dep, "~", paste(rhs, collapse = " + ")))
           
           # --- Fit the penalized logistic regression model ---
-          mod <- glm(fml, data = df, family = binomial(), method = "brglmFit")
+          res <- tryCatch({                  ## added tryCatch
+            mod <- glm(fml, data = df, family = binomial(), method = "brglmFit")
+            broom::tidy(mod, conf.int = TRUE)
+          }, error = function(e) NULL)
+          
+          validate(need(!is.null(res), ""))
           
           # --- Extract tidy summary and check for problematic variables ---
           b <- broom::tidy(mod, conf.int = TRUE)
@@ -7575,8 +7601,8 @@ lmm_vars <- reactive({
         # Check exactly two unique non-NA values
         if (length(dep_levels) != 2) {
           showNotification(strong(
-            sprintf("Dependent variable must have exactly 2 levels, found: %s", paste(dep_levels, collapse = ", "))
-          ), type = "error")
+            sprintf("More than 2 levels were detected. Please use multinomial regressios instead")), 
+            type = "error")
           return(NULL)
         }
         
@@ -7872,7 +7898,8 @@ lmm_vars <- reactive({
       output$nb_dep_ui <- renderUI({
         df <- nb_data()
         count_like_vars <- names(df)[sapply(df, function(col) {
-          is.numeric(col) && all(col >= 0, na.rm = TRUE) && all(col == floor(col), na.rm = TRUE)
+          is.numeric(col) && all(col >= 0, na.rm = TRUE) && all(col == floor(col), na.rm = TRUE) &&
+            length(unique(col[!is.na(col)])) > 5    ## added: must have >5 unique values
         })]
 
         tagList(
@@ -8286,7 +8313,8 @@ lmm_vars <- reactive({
       # ---- Multinomial Regression: Reactive Data ----
       multi_data <- reactive({
         req(input$multi_file)
[…]
