+++ inst/app/app.R
@@ -3042,7 +3042,16 @@ run_wilcoxon_signed_test <- function(df) {
       },
       posthoc_ui,                       # Post hoc button or warning message
       uiOutput("posthoc_result_ui"),    # Then post hoc results
-      actionButton("plot_boxplot_welch_anova", "Plot Boxplot", class = "btn-plot", style = "margin-top: 12px;"),
+      
+      tagList(actionButton("plot_boxplot_welch_anova", "Plot Boxplot", class = "btn-plot", style = "margin-top: 12px;")),
+              # Tooltip icon placed next to the button
+              tags$span(
+                icon("info-circle", class = "fa-solid"),
+                `data-bs-toggle` = "tooltip",
+                `data-bs-placement` = "right",
+                title = "****: p < 0.0001\n***: p < 0.001\n**: p < 0.01\n*: p < 0.05\nns: Not significant",
+                style = "color: #010001; cursor: pointer; margin-left: 6px; position: relative; top: 7px;"
+              ),
       plotOutput("boxplot_welch_anova", height = "350px"),
       downloadButton("download_boxplot_welch_anova", "Download Boxplot", class = "no-print", style = "margin-top: 10px;")
     )
@@ -3265,6 +3274,14 @@ output$welch_boxplot_ui <- renderUI({
   req(welch_clicked())
   tagList(
     actionButton("plot_boxplot_welch", "Plot Boxplot", class = "btn-plot", style = "margin-top: 10px;"),
+    # Tooltip icon placed next to the button
+    tags$span(
+      icon("info-circle", class = "fa-solid"),
+      `data-bs-toggle` = "tooltip",
+      `data-bs-placement` = "right",
+      title = "****: p < 0.0001\n***: p < 0.001\n**: p < 0.01\n*: p < 0.05\nns: Not significant",
+      style = "color: #010001; cursor: pointer; margin-left: 6px; position: relative; top: 5px;"
+    ),
     plotOutput("boxplot_welch", height = "350px"),
     downloadButton("download_boxplot_welch", "Download Boxplot", class = "no-print", style = "margin-top: 10px;")
   )
@@ -3294,7 +3311,21 @@ output$boxplot_ui <- renderUI({
   if (test == "kruskal" && nlevels(df$group) < 3) return(NULL)
   
   tagList(
-    actionButton("plot_boxplot", "Plot Boxplot", class = "btn-plot"),
+    # Action button
+    actionButton(
+      "plot_boxplot",
+      "Plot Boxplot",
+      class = "btn-plot"
+    ),
+    
+    # Tooltip icon placed next to the button
+    tags$span(
+      icon("info-circle", class = "fa-solid"),
+      `data-bs-toggle` = "tooltip",
+      `data-bs-placement` = "right",
+      title = "****: p < 0.0001\n***: p < 0.001\n**: p < 0.01\n*: p < 0.05\nns: Not significant",
+      style = "color: #010001; cursor: pointer; margin-left: 6px;"
+    ),
     plotOutput("boxplot"),
     downloadButton("download_boxplot", "Download Boxplot", class = "no-print")
   )
@@ -3311,8 +3342,9 @@ output$boxplot_ui <- renderUI({
     levs <- levels(df$group)
     # Only allow exactly 2 groups for this test
     if(length(levs) != 2) return(ggplot2::ggplot() + labs(title = "Requires exactly 2 groups"))
+    
     # Explicitly set the comparison
-    #comparison <- list(c(levs[1], levs[2]))
+    comparison <- list(c(levs[1], levs[2]))
     
     ggplot2::ggplot(df, aes(x = group, y = value, fill = group)) +
[…]
