+++ DESCRIPTION
@@ -39,7 +39,7 @@ Imports:
     glmmTMB,
     effectsize,
     coin
-Depends: R (>= 4.0.0)
+Depends: R (>= 4.2.0)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
