+++ inst/app/app.R
@@ -1374,7 +1374,7 @@ server <- function(input, output, session) {
         return()
       }
       if (ngroups > 2) {
-        showNotification(strong("More than two groups detected. Please use One-way ANOVA. Independent t-test is only for comparing exactly two groups."), type = "error")
+        showNotification(strong("More than two groups detected. Please use One-way ANOVA. Independent t-test is only for comparing exactly two groups."), type = "error", duration = 9)
         return()
       }
       if (any(group_sizes < 2)) {
@@ -1392,7 +1392,7 @@ server <- function(input, output, session) {
         return()
       }
       if (ngroups > 2) {
-        showNotification(strong("More than two groups detected. Please use One-way ANOVA. Paired t-test is only for comparing exactly two groups."), type = "error", duration = 5)
+        showNotification(strong("More than two groups detected. Please use One-way ANOVA. Paired t-test is only for comparing exactly two groups."), type = "error", duration = 9)
         return()
       }
       if (length(unique(group_sizes)) > 1) {
@@ -1408,10 +1408,28 @@ server <- function(input, output, session) {
     
     
     # ---- ANOVA GROUP CHECK ----
-    if (input$test_type == "anova" && ngroups < 3) {
-      showNotification(strong("Only two groups detected. Please use a t-test (independent or paired, as appropriate). One-way ANOVA is for three or more groups."), type = "error", duration = 8)
-      return()
+    if (input$test_type == "anova") {
+      ngroups <- nlevels(df$group)
+      
+      if (ngroups == 1) {
+        showNotification(
+          strong("Only one group detected. One-way ANOVA requires three or more groups."), 
+          type = "error", 
+          duration = 9
+        )
+        return()
+      }
+      
+      if (ngroups == 2) {
+        showNotification(
+          strong("Only two groups detected. Please use a t-test (independent or paired, as appropriate). One-way ANOVA is for three or more groups."), 
+          type = "error", 
+          duration = 9
+        )
+        return()
+      }
     }
+    
     processed_data(df)
   })
 
@@ -2440,8 +2458,19 @@ output$levene_text <- renderUI({
 ## Mann-Whitney U test
   run_mannwhitney_test <- function(df) {
     ngroups <- nlevels(df$group)
-    if (ngroups != 2) {
-      showNotification(strong("Mann-Whitney U test requires exactly 2 groups. For more than 2 groups, use Kruskal-Wallis."), type = "error")
+    if (ngroups == 1) {
+      showNotification(
+        strong("Only one group detected. Mann–Whitney U test requires exactly two groups."), 
+        type = "error", duration = 9,
+      )
+      return(data.frame(p = NA))
+    }
+    
+    if (ngroups > 2) {
+      showNotification(
+        strong("More than two groups detected. Check ANOVA assumptions first; if assumptions are violated, AssumpSure will guide you."), 
+        type = "error", duration = 9,
+      )
       return(data.frame(p = NA))
     }
     tryCatch({
@@ -2635,14 +2664,52 @@ run_wilcoxon_signed_test <- function(df) {
     # Show this message on every button click, not inside renderUI!
     if (input$test_type == "wilcoxon_signed") {
       ngroups <- nlevels(df$group)
-      if (ngroups != 2) {
-        showNotification(strong("Wilcoxon signed-rank test requires exactly 2 groups. For more than 2 groups, use Kruskal–Wallis."), type = "error", duration = 9)
+      
+      if (ngroups == 1) {
+        showNotification(
[…]
