+++ inst/app/app.R
@@ -2455,31 +2455,44 @@ output$levene_text <- renderUI({
     }, error = function(e) data.frame(p = NA))
   }
 
-## Mann-Whitney U test
+# # Mann-Whitney U test
+# run_mannwhitney_test <- function(df) {
+#   # ngroups <- nlevels(df$group)
+#   # if (ngroups == 1) {
+#   #   showNotification(
+#   #     strong("Only one group detected. Mann–Whitney U test requires exactly two groups."),
+#   #     type = "error", duration = 9
+#   #   )
+#   #   return(data.frame(p = NA))
+#   # }
+# 
+#   # if (ngroups > 2) {
+#   #   showNotification(
+#   #     strong("More than two groups detected. Check ANOVA assumptions first; if assumptions are violated, AssumpSure will guide you."),
+#   #     type = "error", duration = 9,
+#   #   )
+#   #   return(data.frame(p = NA))
+#   # }
+#   tryCatch({
+#     rstatix::wilcox_test(df, value ~ group, paired = F, detailed = T) %>%
+#       dplyr::mutate(method = "Mann-Whitney")
+#   }, error = function(e) data.frame(p = NA))
+# }
+ 
+   
+
+  ## Mann-Whitney U test
   run_mannwhitney_test <- function(df) {
     ngroups <- nlevels(df$group)
-    if (ngroups == 1) {
-      showNotification(
-        strong("Only one group detected. Mann–Whitney U test requires exactly two groups."), 
-        type = "error", duration = 9,
-      )
-      return(data.frame(p = NA))
-    }
+    if (ngroups != 2) return(data.frame(p = NA))
     
-    if (ngroups > 2) {
-      showNotification(
-        strong("More than two groups detected. Check ANOVA assumptions first; if assumptions are violated, AssumpSure will guide you."), 
-        type = "error", duration = 9,
-      )
-      return(data.frame(p = NA))
-    }
     tryCatch({
-      rstatix::wilcox_test(df, value ~ group, paired = F, detailed = T) %>% 
-        dplyr::mutate(method = "Mann-Whitney")
+      rstatix::wilcox_test(df, value ~ group, paired = FALSE, detailed = TRUE) %>%
+        dplyr::mutate(method = "Mann–Whitney")
     }, error = function(e) data.frame(p = NA))
   }
   
-
+  
 ## Wilcoxon signed-rank test
 run_wilcoxon_signed_test <- function(df) {
   ngroups <- nlevels(df$group)
@@ -2489,7 +2502,8 @@ run_wilcoxon_signed_test <- function(df) {
   if (length(unique(group_sizes)) != 1) return(data.frame(p = NA))
 
   tryCatch({
-    rstatix::wilcox_test(df, value ~ group, paired = TRUE, detailed = TRUE)
+    rstatix::wilcox_test(df, value ~ group, paired = TRUE, detailed = TRUE) %>% 
+      dplyr::mutate(method = "Mann-Whitney")
   }, error = function(e) data.frame(p = NA))
 }
   
@@ -2660,8 +2674,31 @@ run_wilcoxon_signed_test <- function(df) {
     run_test_clicked(Sys.time())
     
     df <- processed_data()
+
+    # Show this message for Mann-Whitney on every button click, not inside renderUI!
+    if (input$test_type == "mannwhitney") {
+      ngroups <- nlevels(df$group)
+      
+      if (ngroups == 1) {
+        showNotification(
+          strong("Only one group detected. Mann–Whitney U test requires exactly two groups."), 
+          type = "error", 
+          duration = 9
+        )
+        return()
+      }
+      
+      if (ngroups > 2) {
+        showNotification(
[…]
