+++ R/zzz.R
@@ -1,6 +1,6 @@
 .onAttach <- function(libname, pkgname) {
   current <- utils::packageVersion("AssumpSure")
-  latest  <- "1.1.1"   # hard-code my current CRAN release
+  latest  <- "1.1.2"   # hard-code my current CRAN release
   
   if (current < latest) {
     packageStartupMessage(
