+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: AssumpSure
 Title: 'shiny' Application for Statistical Test Assumption Checking and Guidance
-Version: 1.1.1
+Version: 1.1.2
 Authors@R: 
     person("Ahmed", "Bargheet", , "ahmed.bargheet@yahoo.com", role = c("aut", "cre"))
 Description: A 'shiny' application to assess statistical assumptions and guide users toward appropriate tests. The app is designed for researchers with minimal statistical training and provides diagnostics, plots, and test recommendations for a wide range of analyses. Many statistical assumptions are implemented using the package 'rstatix' (Kassambara, 2019) <doi:10.32614/CRAN.package.rstatix> and 'performance' (Lüdecke et al., 2021) <doi:10.21105/joss.03139>.
