### 2025-10-18 5284013
README.md
Added a badge for last month's CRAN downloads.
