+++ README.md
@@ -2,6 +2,7 @@
 
 [![](https://www.r-pkg.org/badges/version/AssumpSure?color=orange)](https://cran.r-project.org/package=AssumpSure) ![License: GPL-3](https://img.shields.io/badge/license-GPL--3-blue.svg) [![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![CRAN downloads](https://cranlogs.r-pkg.org/badges/grand-total/AssumpSure)](https://cran.r-project.org/package=AssumpSure)
+[![CRAN downloads](https://cranlogs.r-pkg.org/badges/last-month/AssumpSure)](https://cran.r-project.org/package=AssumpSure)
 
 
 ## 🎯 Purpose
