### 2025-10-18 28d45c7
README.md
Removed CRAN download badges from README.
