+++ inst/app/app.R
@@ -6262,6 +6262,15 @@ output$cor_matrix_download_ui <- renderUI({
         rhs <- c(main_effects, interaction_terms)
         fml <- as.formula(paste(input$lm_dep, "~", paste(rhs, collapse = " + ")))
         
+        # >>> INSERTED: fit model and compute fit metrics
+        mod <- lm(fml, data = df)
+        smry <- summary(mod)
+        r2   <- unname(smry$r.squared)
+        adjr <- unname(smry$adj.r.squared)
+        aic  <- AIC(mod)
+        bic  <- BIC(mod)
+        rmse <- sqrt(mean(residuals(mod)^2))
+        
         tbl_lm <-lm(fml, data = df) %>%
           broom::tidy(conf.int = T) %>%
           dplyr::mutate(Sig = dplyr::case_when(
@@ -6276,8 +6285,39 @@ output$cor_matrix_download_ui <- renderUI({
             bootstrap_options = c("hover", "condensed", "bordered"),
             position = "center", full_width = T) %>% 
           kableExtra::scroll_box(width = "100%", height = "100%")
-        div(style = "margin-top: 15px;",
-        HTML(tbl_lm)
+        
+        # >>> INSERTED: model-fit square (heading + quick guidance)
+        fit_box <- div(
+          style = "margin-top: 14px; background:#f5f7fb; border:1px solid #d9e1f2; border-radius:8px; padding:12px;",
+          div(
+            style = "display:flex; align-items:center; gap:8px; margin-bottom:8px;",
+            tags$strong("Model fit & what to do"),
+            tags$span(
+              icon("info-circle", class = "fa-solid"),
+              `data-bs-toggle` = "tooltip",
+              `data-bs-placement` = "right",
+              title = "Akaike Information Criterion (AIC)\nBayesian Information Criterion (BIC)\nRoot Mean Square Error (RMSE)\nAdjusted R-squared (Adj. R²)\n\n
+Compare models fitted on the same data:\n
+• Lower AIC, BIC, and RMSE indicate better fit.\n• Higher adjusted R² indicates better explanatory power.\n
+If AIC differs by less than 2, the models fit equally well, so prefer the simpler one.",
+              class = "no-print",
+              style = "color:black; cursor:pointer;"
+            )
+          ),
+          tags$div(
+            style = "display:grid; grid-template-columns: repeat(5, minmax(0,1fr)); gap:8px; font-size:0.95em;",
+            tags$div(tags$strong("AIC"),  br(), sprintf("%.2f", aic)),
+            tags$div(tags$strong("BIC"),  br(), sprintf("%.2f", bic)),
+            tags$div(tags$strong("R²"),   br(), sprintf("%.3f", r2)),
+            tags$div(tags$strong("Adj. R²"), br(), sprintf("%.3f", adjr)),
+            tags$div(tags$strong("RMSE"), br(), sprintf("%.3f", rmse))
+          )
+        )
+        
+        div(
+          style = "margin-top: 15px;",
+          HTML(tbl_lm),
+          fit_box
         )
       })
     })
@@ -6296,9 +6336,12 @@ output$cor_matrix_download_ui <- renderUI({
               icon("info-circle", class = "fa-solid"),
               `data-bs-toggle` = "tooltip",
               `data-bs-placement` = "right",
-              title = "• If residuals are not normally distributed or variance is not constant (e.g., residuals vs. fitted plot shows a funnel shape), consider transforming the dependent variable (e.g., log, Box-Cox, or Yeo-Johnson).\n
-              • If VIF > 10, remove or combine correlated predictors.\n
-              • Check for influential outliers and consider removing or justifying them.",
[…]
