### 2025-10-22 #9286 @editorialbot
[REVIEW]: AssumpSure: a user-friendly R Shiny package for automated validation of statistical assumptions and appropriate test selection
**Submitting author:** @Ahmedbargheet (<a href="http://orcid.org/0000-0002-0363-0249">Ahmed Bargheet</a>)
**Repository:** https://github.com/Ahmedbargheet/AssumpSure
**Branch with paper.md** (empty if default branch): paper
**Version:** v.1.1.0
