### 2025-10-22 #9286 @adithirgis
Hello @Ahmedbargheet,

Congratulations on this app. Looks interesting! I have started the review process.

I had one query - I am unable to see Community Guidelines in the Documentation. Please let me know if I have missed it. Something similar to this `usethis::use_tidy_contributing()`

Thanks,
Adithi
