### 2025-10-22 #9286 @adithirgis
Hello,

I appreciate the quick response. I see that when I install `AssumpSure` and then `launch_app()` when I try to Download the Sample Data it takes some time to download the data infants.csv is that correct?

While when I download from the online version [here](https://ahmed-bargheet.shinyapps.io/AssumpSure/) I land on the GitHub repo csv location - https://github.com/Ahmedbargheet/AssumpSure/blob/main/inst/extdata/infants.csv. Can this be automatically downloaded (in case it is difficult to download from GitHub)?

Thanks,
Adithi
