### 2025-10-22 #9286 @adithirgis
Hi @Ahmedbargheet,

I would request you to add a tests folder with automated tests for your functions using either [testthat](https://testthat.r-lib.org/) or [shinytest2](https://github.com/rstudio/shinytest2).

`Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
`
Thanks,
Adithi
