+++ tests/testthat/test-basic.R
@@ -1,0 +1,3 @@
+test_that("App launches without error", {
+  expect_error(AssumpSure::launch_app(), NA)
+})
