+++ man/sample_data.Rd
@@ -1,0 +1,39 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sample_data.R
+\name{sample_data}
+\alias{sample_data}
+\title{Sample Infant Dataset (CSV)}
+\format{
+A CSV file with 251 rows and 12 columns:
+\describe{
+\item{infant_id}{Participant identifier.}
+\item{Treatment}{Treatment group.}
+\item{timepoint}{Measurement time point.}
+\item{gender}{Participant gender.}
+\item{born_method}{Method of delivery.}
+\item{country}{Participant country.}
+\item{fortified}{Whether the participant received fortification.}
+\item{nutrients}{Whether the participant received nutrients.}
+\item{diarrhea}{Whether the participant experienced diarrhea.}
+\item{weight}{Participant weight.}
+\item{count}{Bacterial richness count.}
+\item{height}{Participant height.}
+}
+}
+\source{
+Simulated data generated by the author.
+}
+\description{
+A simulated dataset included with the AssumpSure package and used by the Shiny app
+for demonstration and testing.
+}
+\details{
+The file is stored in \code{inst/extdata/infants.csv} and can be accessed with:
+\code{system.file("extdata", "infants.csv", package = "AssumpSure")}.
+
+All values are synthetic, created only for demonstration; no human data were used,
+and no ethical approval or consent is required.
+}
+\examples{
+read.csv(system.file("extdata", "infants.csv", package = "AssumpSure")))
+}
