+++ DESCRIPTION
@@ -46,3 +46,6 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 URL: https://github.com/Ahmedbargheet/AssumpSure
 BugReports: https://github.com/Ahmedbargheet/AssumpSure/issues
+Suggests: 
+    testthat (>= 3.0.0)
+Config/testthat/edition: 3
