+++ man/sample_data.Rd
@@ -19,6 +19,11 @@ A CSV file with 251 rows and 12 columns:
 \item{count}{Bacterial richness count.}
 \item{height}{Participant height.}
 }
+
+Sample Datasets in AssumpSure
+
+AssumpSure includes several simulated datasets to demonstrate different
+statistical tests within the Shiny app.
 }
 \source{
 Simulated data generated by the author.
@@ -34,6 +39,21 @@ The file is stored in \code{inst/extdata/infants.csv} and can be accessed with:
 All values are synthetic, created only for demonstration; no human data were used,
 and no ethical approval or consent is required.
 }
+\section{Available datasets}{
+
+\describe{
+\item{infants.csv}{Used for continuous data tests, Fisher & Chi-square, and regression models.}
+\item{t-test.csv}{Used for independent and paired t-tests.}
+\item{plantgrowth.csv}{Used for for one-way ANOVA.}
+\item{chi_fisher.csv}{Used for Fisher & Chi-square.}
+\item{correlation.csv}{Used for correlation.}
+\item{bacteria_for_correlation.csv}{Used for correlation for compositional data.}
+}
+
+Each dataset is synthetic and created only for demonstration. No human data
+were used, and no ethical approval or consent is required.
+}
+
 \examples{
-read.csv(system.file("extdata", "infants.csv", package = "AssumpSure")))
+read.csv(system.file("extdata", "infants.csv", package = "AssumpSure"))
 }
