### 2025-10-24 bc7785f
paper.md
