+++ paper/paper.md
@@ -24,29 +24,29 @@ link-citations: yes
 
 # Summary
 
-`AssumpSure` is an open-source R Shiny package that enforces mandatory assumption checks to improve statistical rigor and guide appropriate test selection. It supports a wide array of statistical tests. The app automatically evaluates key assumptions such as normality (Shapiro–Wilk) and homogeneity of variance (Levene’s test), using color-coded indicators (green for passed, red for violated) with clear interpretive summaries. When assumptions are not met, it recommends alternatives like Kruskal–Wallis for non-normal ANOVA data. `AssumpSure` also prevents invalid variable types, provides preprocessing tools (log, Box–Cox, centered log-ratio transformations), and generates downloadable diagnostic plots and publication-ready outputs.
+*AssumpSure* is an open-source R Shiny package that enforces mandatory assumption checks to improve statistical rigor and guide appropriate test selection. It supports a wide array of statistical tests. The app automatically evaluates key assumptions such as normality (Shapiro–Wilk) and homogeneity of variance (Levene’s test), using color-coded indicators (green for passed, red for violated) with clear interpretive summaries. When assumptions are not met, it recommends alternatives like Kruskal–Wallis for non-normal ANOVA data. *AssumpSure* also prevents invalid variable types, provides preprocessing tools (log, Box–Cox, centered log-ratio transformations), and generates downloadable diagnostic plots and publication-ready outputs.
 
 # Statement of Need
 Rigorous statistical analysis depends on checking underlying assumptions and choosing appropriate tests. Yet in practice, these steps are often skipped, misunderstood, or incorrectly applied. Reviews of social science literature report that fewer than 25% of studies explicitly mention assumption checks, and even fewer report them correctly [@shatz2024assumption]. This contributes to widespread misuse of statistical tests and undermines reproducibility. One common issue is the misinterpretation of assumption tests. For example, many users mistakenly interpret a significant Shapiro–Wilk test result (*p* < 0.05) as evidence for normality rather than deviation from it [@ghasemi2012normality; @midway2025testing; @gosselin2024testing]. Moreover, in small samples, where the Shapiro–Wilk test lacks power, borderline results are easily overlooked or misjudged.
 
-`AssumpSure` directly addresses these problems by making assumption checking a required part of the analysis workflow. Through an intuitive web interface, users are guided step-by-step: assumptions are tested automatically, results are color-coded (green for met, red for violated, orange for borderline), and clear recommendations are given. For example, if ANOVA assumptions are not met, Kruskal–Wallis is suggested. By complementing color-coded squares with tooltips and clear messages, `AssumpSure` helps reduce misinterpretation of common assumption tests and guides the user on how to proceed.
+*AssumpSure* directly addresses these problems by making assumption checking a required part of the analysis workflow. Through an intuitive web interface, users are guided step-by-step: assumptions are tested automatically, results are color-coded (green for met, red for violated, orange for borderline), and clear recommendations are given. For example, if ANOVA assumptions are not met, Kruskal–Wallis is suggested. By complementing color-coded squares with tooltips and clear messages, *AssumpSure* helps reduce misinterpretation of common assumption tests and guides the user on how to proceed.
 
 # Comparison to other Packages
[…]
