+++ R/infants.R
@@ -25,27 +25,9 @@
 #'   \item{height}{Participant height.}
 #' }
 #'
-#' Sample Datasets in AssumpSure
-#'
-#' AssumpSure includes several simulated datasets to demonstrate different
-#' statistical tests within the Shiny app.
-#'
-#' @section Available datasets:
-#' \describe{
-#'   \item{infants.csv}{Used for continuous data tests, Fisher & Chi-square, and regression models.}
-#'   \item{t-test.csv}{Used for independent and paired t-tests.}
-#'   \item{plantgrowth.csv}{Used for for one-way ANOVA.}
-#'   \item{chi_fisher.csv}{Used for Fisher & Chi-square.}
-#'   \item{correlation.csv}{Used for correlation.}
-#'   \item{bacteria_for_correlation.csv}{Used for correlation for compositional data.}
-#' }
-#'
-#' Each dataset is synthetic and created only for demonstration. No human data
-#' were used, and no ethical approval or consent is required.
-#'
 #' @source Simulated data generated by the author.
 #'
 #' @examples
 #' read.csv(system.file("extdata", "infants.csv", package = "AssumpSure"))
-#' @name sample_data
+#' @name infants
 NULL
