+++ man/infants.Rd
@@ -1,7 +1,7 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/sample_data.R
-\name{sample_data}
-\alias{sample_data}
+% Please edit documentation in R/infants.R
+\name{infants}
+\alias{infants}
 \title{Sample Infant Dataset (CSV)}
 \format{
 A CSV file with 251 rows and 12 columns:
@@ -19,11 +19,6 @@ A CSV file with 251 rows and 12 columns:
 \item{count}{Bacterial richness count.}
 \item{height}{Participant height.}
 }
-
-Sample Datasets in AssumpSure
-
-AssumpSure includes several simulated datasets to demonstrate different
-statistical tests within the Shiny app.
 }
 \source{
 Simulated data generated by the author.
@@ -39,21 +34,6 @@ The file is stored in \code{inst/extdata/infants.csv} and can be accessed with:
 All values are synthetic, created only for demonstration; no human data were used,
 and no ethical approval or consent is required.
 }
-\section{Available datasets}{
-
-\describe{
-\item{infants.csv}{Used for continuous data tests, Fisher & Chi-square, and regression models.}
-\item{t-test.csv}{Used for independent and paired t-tests.}
-\item{plantgrowth.csv}{Used for for one-way ANOVA.}
-\item{chi_fisher.csv}{Used for Fisher & Chi-square.}
-\item{correlation.csv}{Used for correlation.}
-\item{bacteria_for_correlation.csv}{Used for correlation for compositional data.}
-}
-
-Each dataset is synthetic and created only for demonstration. No human data
-were used, and no ethical approval or consent is required.
-}
-
 \examples{
 read.csv(system.file("extdata", "infants.csv", package = "AssumpSure"))
 }
