### 2025-10-24 #9286 @Ahmedbargheet
Hello @adithirgis,

Thank you for the detailed feedback.

1. **Tests:** Additional automated tests have been added to ensure representative coverage of all statistical modules in *AssumpSure*. New tests now include Fisher’s exact test, correlation, and linear regression, complementing the existing continuous-data tests. Together, these cover the main analytical functions available in the app, and all tests now pass successfully.

2. **Example usage:** Added a new section titled **Example Usage and Reproducibility** in the manuscript, providing example datasets and clear usage instructions for each tab.

3. **Dataset documentation:** Split the previous combined sample_data documentation into separate help files for each dataset (`infants.csv,` `t-test.csv`, `plantgrowth.csv`, `chi_fisher.csv`, `correlation.csv`, and `bacteria_for_correlation.csv`), each with its own description and example.

4. **Session termination:** I was unable to reproduce the reported issue when pressing Esc/Stop. However, to ensure consistent behavior across environments, I added safeguards in the server code using `session$onSessionEnded(stopApp()) `and a `safe_run()` wrapper to handle user interrupts gracefully.

Thank you again for the constructive review.

Best,
Ahmed
