+++ README.md
@@ -109,7 +109,77 @@ If you use **AssumpSure** in your research, please cite it as below.
 Bargheet, Ahmed. "AssumpSure: A User-Friendly R Shiny Package for Automated Validation of Statistical Assumptions and Appropriate Test Selection." Available at SSRN: https://ssrn.com/abstract=5352013 or http://dx.doi.org/10.2139/ssrn.5352013.
 
 ---
+## 📚 Reproducible Examples
+The examples below demonstrate how to use each analysis module in *AssumpSure*.
+All example datasets are included [here](https://github.com/Ahmedbargheet/AssumpSure/tree/main/inst/extdata) or downloadable within the app (Help tab).
+Screenshots display the expected outputs for verifying functionality and reproducibility.
 
+**1. Independent t-test**
+**Dataset:** `t-test.csv`
+**Steps:**
+- Launch the app:
+```r
+AssumpSure::launch_app()
+```
+- Upload `t-test.csv` in the **Continuous Data Tests** tab.
+- Choose **Independent T-test**.
+- Select the numeric value and `Treatment` as a categorical value, and click **Check Assumptions**.
+  
+![t-test example](inst/app/www/independent_t.test.png)
+
+**2. One-way ANOVA**
+**Dataset:** `plantgrowth.csv`
+- Upload `plantgrowth.csv` in the **Continuous Data Tests** tab.
+- Choose **One-way ANOVA**.
+- Choose the numeric value and the categorical value, and click **Check Assumptions**.
+- Then click on **Run Test**, and choose **p-value adjustment method or Tukey HSD**, and click on **Run the Selected Option**
+- Optionally, click **Plot Boxplot** to visualize group differences.
+  
+![One-way_ANOVA example](inst/app/www/One-way_ANOVA.png)
+![ANOVA_test example](inst/app/www/ANOVA_test.png)
+![ANOVA_test_1 example](inst/app/www/ANOVA_test_1.png)
+
+
+**3. Fisher's exact test**
+**Dataset:** `chi_fisher.csv`
+- Upload `chi_fisher.csv` in the **Fisher & Chi-square** tab.
+- Choose **sex** as first categorical variable and **location** as second categorical variable.
+- Choose **Fisher's exact test** from the **Choose test** drop menu and click on **Run Test**.
+- Choose **Benjamini-Hochberg**
+- Optionally, click **Plot** to visualize groups.
+  
+![fisher example](inst/app/www/fisher.png)
+
+
+**3. Linear Model**
+**Dataset:** `infants.csv`
+- Upload `infants.csv` in the **Regression Models** tab.
+- Choose **weight** as dependent variable and **Treatment, gender, country, nutrients, and diarrhea** as independent variables.
+- Choose **Inverse normal** from the **Transform dependent variable** drop menu and click on **Check Assumptions**.
+- Then click on **run LM** and **Plot Forest**.
+  
+![LM example](inst/app/www/LM.png)
+![LM2 example](inst/app/www/LM2.png)
+![LM3 example](inst/app/www/LM3.png)
+
+
+**4. Correlation**
+**Dataset:** `correlation.csv`
+- Upload `correlation.csv` in the **Correlation** tab.
+- Choose **Pearson** from the **Correlation Method** drop menu and click **Check Assumptions**.
+  
+![correlation example](inst/app/www/correlation.png)
+
+
+**5. Correlation for compositional data**
+**Dataset:** `bacteria_for_correlation.csv`
+- Upload `bacteria_for_correlation.csv` in the **Correlation** tab.
+- Tick **Apply CLR transformation**, and set **Remove features** to **20%** using the slider
+- Choose **Spearman** from the **Correlation Method** drop menu and click **Run Correlation**.
[…]
