+++ README.md
@@ -114,9 +114,9 @@ The examples below demonstrate how to use each analysis module in *AssumpSure*.
 All example datasets are included [here](https://github.com/Ahmedbargheet/AssumpSure/tree/main/inst/extdata) or downloadable within the app (Help tab).
 Screenshots display the expected outputs for verifying functionality and reproducibility.
 
-**1. Independent t-test**
-**Dataset:** `t-test.csv`
-**Steps:**
+**1. Independent t-test**  
+**Dataset:** `t-test.csv`  
+**Steps:**  
 - Launch the app:
 ```r
 AssumpSure::launch_app()
@@ -127,9 +127,9 @@ AssumpSure::launch_app()
   
 ![t-test example](inst/app/www/independent_t.test.png)
 
-**2. One-way ANOVA**
-**Dataset:** `plantgrowth.csv`
-- Upload `plantgrowth.csv` in the **Continuous Data Tests** tab.
+**2. One-way ANOVA**  
+**Dataset:** `plantgrowth.csv`  
+- Upload `plantgrowth.csv` in the **Continuous Data Tests** tab.  
 - Choose **One-way ANOVA**.
 - Choose the numeric value and the categorical value, and click **Check Assumptions**.
 - Then click on **Run Test**, and choose **p-value adjustment method or Tukey HSD**, and click on **Run the Selected Option**
@@ -140,8 +140,8 @@ AssumpSure::launch_app()
 ![ANOVA_test_1 example](inst/app/www/ANOVA_test_1.png)
 
 
-**3. Fisher's exact test**
-**Dataset:** `chi_fisher.csv`
+**3. Fisher's exact test**  
+**Dataset:** `chi_fisher.csv`  
 - Upload `chi_fisher.csv` in the **Fisher & Chi-square** tab.
 - Choose **sex** as first categorical variable and **location** as second categorical variable.
 - Choose **Fisher's exact test** from the **Choose test** drop menu and click on **Run Test**.
@@ -151,8 +151,8 @@ AssumpSure::launch_app()
 ![fisher example](inst/app/www/fisher.png)
 
 
-**3. Linear Model**
-**Dataset:** `infants.csv`
+**3. Linear Model**  
+**Dataset:** `infants.csv`  
 - Upload `infants.csv` in the **Regression Models** tab.
 - Choose **weight** as dependent variable and **Treatment, gender, country, nutrients, and diarrhea** as independent variables.
 - Choose **Inverse normal** from the **Transform dependent variable** drop menu and click on **Check Assumptions**.
@@ -163,16 +163,16 @@ AssumpSure::launch_app()
 ![LM3 example](inst/app/www/LM3.png)
 
 
-**4. Correlation**
-**Dataset:** `correlation.csv`
+**4. Correlation**  
+**Dataset:** `correlation.csv`  
 - Upload `correlation.csv` in the **Correlation** tab.
 - Choose **Pearson** from the **Correlation Method** drop menu and click **Check Assumptions**.
   
 ![correlation example](inst/app/www/correlation.png)
 
 
-**5. Correlation for compositional data**
-**Dataset:** `bacteria_for_correlation.csv`
+**5. Correlation for compositional data**  
+**Dataset:** `bacteria_for_correlation.csv`  
 - Upload `bacteria_for_correlation.csv` in the **Correlation** tab.
 - Tick **Apply CLR transformation**, and set **Remove features** to **20%** using the slider
 - Choose **Spearman** from the **Correlation Method** drop menu and click **Run Correlation**.
