### 2025-10-24 67bb81a
paper.md
Updated illustrative example to specify the use of the 'infants.csv' dataset with AssumpSure.
