### 2025-10-24 93c66e0
README.md
Updated images in README with new filenames.
