### 2025-10-24 #9286 @Ahmedbargheet
Hello again @adithirgis,

By the way, I realized that different versions of the `infants.csv` dataset were previously used in the manuscript and the `inst/extdata` folder. The correct dataset has now been uploaded to `inst/extdata`, ensuring full reproducibility between the paper and the app. This correction has also been applied to GitHub and the online versions. The CRAN version will be updated once the JOSS review process is complete.

Best,
Ahmed
