### 2025-10-25 #9286 @adithirgis
Hello @Ahmedbargheet,

I have completed the review process. All the best for the app. The app looks good and can be used for various statistical tests.

@crvernon I have now completed the review now.

Thanks,
Adithi
