+++ DESCRIPTION
@@ -5,42 +5,42 @@ Authors@R:
     person("Ahmed", "Bargheet", , "ahmed.bargheet@yahoo.com", role = c("aut", "cre"))
 Description: A 'shiny' application to assess statistical assumptions and guide users toward appropriate tests. The app is designed for researchers with minimal statistical training and provides diagnostics, plots, and test recommendations for a wide range of analyses. Many statistical assumptions are implemented using the package 'rstatix' (Kassambara, 2019) <doi:10.32614/CRAN.package.rstatix> and 'performance' (Lüdecke et al., 2021) <doi:10.21105/joss.03139>.
 License: GPL-3
-Imports: 
-    shiny (>=1.11.1),
-    shinyjs (>=2.1.0),
-    shinyBS (>=0.61.1),
-    bslib (>=0.9.0),
-    htmltools (>=0.5.8.1),
-    shinyscreenshot (>=0.2.1.9000),
-    fontawesome (>=0.5.3),
-    tidyverse (>=2.0.0),
-    rstatix (>=0.7.2),
-    ggpubr (>=0.6.1),
-    knitr (>=1.50),
-    compositions (>=2.0.8),
-    car (>=3.1.3),
-    correlation (>=0.8.8),
-    MVN (>=6.1),
-    performance (>=0.15.0),
-    modelbased (>=0.12.0),
-    patchwork (>=1.3.1),
-    see (>=0.11.0),
-    bestNormalize (>=1.9.1),
-    DHARMa (>=0.4.7),
-    lmerTest (>=3.1.3),
-    lme4 (>=1.1.37),
-    MASS (>=7.3.65),
-    nnet (>=7.3.20),
-    broom (>=1.0.8),
-    broom.mixed (>=0.2.9.6),
-    DT (>=0.33),
-    kableExtra (>=1.4.0),
-    sjPlot (>=2.9.0),
-    brglm2 (>=0.9.2),
-    glmmTMB (>=1.1.11),
-    effectsize (>=1.0.1),
-    coin (>=1.4.3),
-    dplyr (>=1.1.4)
+Imports:
+    shiny (>= 1.11.1),
+    shinyjs (>= 2.1.0),
+    shinyBS (>= 0.61.1),
+    bslib (>= 0.9.0),
+    htmltools (>= 0.5.8.1),
+    shinyscreenshot (>= 0.2.1.9000),
+    fontawesome (>= 0.5.3),
+    tidyverse (>= 2.0.0),
+    rstatix (>= 0.7.2),
+    ggpubr (>= 0.6.1),
+    knitr (>= 1.50),
+    compositions (>= 2.0.8),
+    car (>= 3.1.3),
+    correlation (>= 0.8.8),
+    MVN (>= 6.1),
+    performance (>= 0.15.0),
+    modelbased (>= 0.12.0),
+    patchwork (>= 1.3.1),
+    see (>= 0.11.0),
+    bestNormalize (>= 1.9.1),
+    DHARMa (>= 0.4.7),
+    lmerTest (>= 3.1.3),
+    lme4 (>= 1.1.37),
+    MASS (>= 7.3.65),
+    nnet (>= 7.3.20),
+    broom (>= 1.0.8),
+    broom.mixed (>= 0.2.9.6),
+    DT (>= 0.33),
+    kableExtra (>= 1.4.0),
+    sjPlot (>= 2.9.0),
+    brglm2 (>= 0.9.2),
+    glmmTMB (>= 1.1.11),
+    effectsize (>= 1.0.1),
+    coin (>= 1.4.3),
+    dplyr (>= 1.1.4)
 Depends: R (>= 4.2.0)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
