### 2025-11-05 #9286 @Ahmedbargheet
Thank you, @ZekeMarshall, for reviewing my submission and for your kind words. I’m really happy that you found my app well-designed and easy to use.

**Answer to comment 1:**
Thank you very much for the recommendation. _AssumpSure_ already follows a modular structure at the functional and analytical levels, as described in the manuscript. The app is launched through the exported `launch_app()` function in [R/launch_app.R](https://github.com/Ahmedbargheet/AssumpSure/blob/main/R/launch_app.R), which encapsulates the full Shiny system and runs it directly from the installed package. Each analytical domain, continuous variable tests, categorical tests, correlations, and regression models, is implemented as a distinct functional unit with its own logic and interface components. While splitting would, of course, improve code readability, it would require substantial restructuring without adding functional or modular benefits. All functions are documented with `roxygen2`, and the app’s core functionality is described in the package documentation. I would sincerely appreciate it if the current organization could be accepted as meeting JOSS standards for modularity and maintainability.

**Answer to comment 2:**
I have added unit tests for the main statistical functions (including `paired t-tests`, `One-way ANOVA`, `Kruskal-Wallis`, `Mann-Whitney`, and `Wilcoxon`) using `rstatix`. These tests are now included in the [tests/testthat/](https://github.com/Ahmedbargheet/AssumpSure/tree/main/tests/testthat) directory and all pass successfully. This approach improves reliability and maintainability while keeping the app’s structure stable. I hope this solution meets the intent of your recommendation and is acceptable as consistent with JOSS requirements.

**Answer to comment 3:**
Thank you for the helpful suggestion. I have now added minimum version requirements for all imported packages in the [DESCRIPTION](https://github.com/Ahmedbargheet/AssumpSure/blob/main/DESCRIPTION) file and initialized `renv` to record exact dependency versions in a lockfile, ensuring full reproducibility. Please see [renv](https://github.com/Ahmedbargheet/AssumpSure/tree/main/renv) and [here](https://github.com/Ahmedbargheet/AssumpSure/blob/main/renv.lock)

Thanks again for the review.

Regards,
Ahmed Bargheet
