### 2025-11-05 712bdb0
paper.bib
Removed unnecessary apostrophe from 'AssumpSure’s' in the Software Description section.
