+++ paper/paper.md
@@ -35,7 +35,7 @@ Rigorous statistical analysis depends on checking underlying assumptions and cho
 Tools like *Jamovi* [@csahin2019jamovi] and *JASP* [@love2019jasp] offer user-friendly interfaces but lack integrated workflows that enforce mandatory assumption diagnostics, appropriate test selection, and actionable recommendations, leaving users to manually interpret complex outputs. This increases errors and undermines reproducibility. Additionally, non-R applications such as assumption-sheriff (Python) [@assumptionsheriff] also aim to simplify statistical assumption checking and test selection. However, these alternatives lack fully integrated, interactive diagnostic workflows or do not provide enforced guidance on test appropriateness for each step.
 
 # Software Description
-*AssumpSure*’s modular R Shiny architecture organizes statistical analyses into continuous variable tests, categorical tests, correlations, and regression models. Users upload CSV data, and the interface automatically detects variable types, preventing invalid selections with clear error messages (e.g., suggesting one-way ANOVA if a t-test is chosen for a categorical variable with three levels). Mandatory assumption diagnostics (normality via Shapiro-Wilk, variance homogeneity via Levene’s test) are visualized through diagnostic plots (QQ plots, histograms, scatter plots), with tooltips to guide the user how to interpret the plots.
+*AssumpSure*’s R Shiny architecture organizes statistical analyses into continuous variable tests, categorical tests, correlations, and regression models. Users upload CSV data, and the interface automatically detects variable types, preventing invalid selections with clear error messages (e.g., suggesting one-way ANOVA if a t-test is chosen for a categorical variable with three levels). Mandatory assumption diagnostics (normality via Shapiro-Wilk, variance homogeneity via Levene’s test) are visualized through diagnostic plots (QQ plots, histograms, scatter plots), with tooltips to guide the user how to interpret the plots.
 
 For correlations, a tooltip explains the centered log-ratio (CLR) transformation’s importance, conducted by compositions R package [@van2013package], in reducing spurious correlations in compositional data (e.g., microbiome profiles). Prevalence filtering is available for correlations to enhance reliability. Regression models leverage the performance package [@ludecke2021performance], providing transformations like log or Box-Cox that are offered to improve residual normality and reduce heteroscedasticity, with guidance on their role in ensuring model validity. 
 
