+++ inst/app/app.R
@@ -4128,31 +4128,31 @@ output$boxplot_ui <- renderUI({
 
 
 
-    stat_square_chisq <- function(df) {
-      res <- run_chisq_test(df)
-      pval <- res$p[1]
-      if (is.na(pval)) return(NULL)
-      div(
-        style = paste0(
-          "background-color: ", if (pval < 0.05) "green" else "#B20D00",
-          "; color: white; padding: 10px; border-radius: 5px; margin-top: 10px;"
-        ),
-        strong(if (pval < 0.05) "Statistically Significant Difference" else "No Statistically Significant Difference")
-      )
-    }
+    # stat_square_chisq <- function(df) {
+    #   res <- run_chisq_test(df)
+    #   pval <- res$p[1]
+    #   if (is.na(pval)) return(NULL)
+    #   div(
+    #     style = paste0(
+    #       "background-color: ", if (pval < 0.05) "green" else "#B20D00",
+    #       "; color: white; padding: 10px; border-radius: 5px; margin-top: 10px;"
+    #     ),
+    #     strong(if (pval < 0.05) "Statistically Significant Difference" else "No Statistically Significant Difference")
+    #   )
+    # }
 
-    stat_square_fisher <- function(df) {
-      res <- run_fisher_test(df)
-      pval <- res$p[1]
-      if (is.na(pval)) return(NULL)
-      div(
-        style = paste0(
-          "background-color: ", if (pval < 0.05) "green" else "#B20D00",
-          "; color: white; padding: 10px; border-radius: 5px; margin-top: 10px;"
-        ),
-        strong(if (pval < 0.05) "Statistically Significant Difference" else "No Statistically Significant Difference")
-      )
-    }
+    # stat_square_fisher <- function(df) {
+    #   res <- run_fisher_test(df)
+    #   pval <- res$p[1]
+    #   if (is.na(pval)) return(NULL)
+    #   div(
+    #     style = paste0(
+    #       "background-color: ", if (pval < 0.05) "green" else "#B20D00",
+    #       "; color: white; padding: 10px; border-radius: 5px; margin-top: 10px;"
+    #     ),
+    #     strong(if (pval < 0.05) "Statistically Significant Difference" else "No Statistically Significant Difference")
+    #   )
+    # }
 
 
     # Handle post hoc
@@ -4527,13 +4527,13 @@ output$boxplot_ui <- renderUI({
   last_feature_count <- reactiveVal(NULL)
   
   # --- PATCH 0:  ---
-  is_singular_matrix <- function(S, tol = .Machine$double.eps^0.5) {
-    if (!is.matrix(S)) return(TRUE)
-    if (any(!is.finite(S))) return(TRUE)
-    r <- qr(S)$rank
-    p <- ncol(S)
-    (r < p) || any(is.na(S)) || (min(abs(eigen(S, symmetric = TRUE, only.values = TRUE)$values)) < tol)
-  }
+  # is_singular_matrix <- function(S, tol = .Machine$double.eps^0.5) {
+  #   if (!is.matrix(S)) return(TRUE)
+  #   if (any(!is.finite(S))) return(TRUE)
+  #   r <- qr(S)$rank
+  #   p <- ncol(S)
+  #   (r < p) || any(is.na(S)) || (min(abs(eigen(S, symmetric = TRUE, only.values = TRUE)$values)) < tol)
+  # }
   
   
 
@@ -4553,9 +4553,9 @@ output$boxplot_ui <- renderUI({
 
 # Data Loading & Numeric Feature Detection
 
-  safe_shapiro_pval <- function(x) {
-    !is.null(x) && !is.null(x$p.value) && is.numeric(x$p.value) && !is.na(x$p.value)
-  }
[…]
