+++ README.md
@@ -4,6 +4,7 @@
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17539689.svg)](https://doi.org/10.5281/zenodo.17539689)
 
 
+
 ## 🎯 Purpose
 <img align="right" src="inst/app/www/logo.png" width="200" style="margin-top:40px;">
 
