### 2025-11-06 b6c785a
Create CITATION.cff
