+++ paper/paper.bib
@@ -83,7 +83,7 @@
 }
 
 @article{ludecke2021performance,
-  title={performance: An R package for assessment, comparison and testing of statistical models},
+  title={performance: An {R} package for assessment, comparison and testing of statistical models},
   author={L{\"u}decke, Daniel and Ben-Shachar, Mattan S and Patil, Indrajeet and Waggoner, Philip and Makowski, Dominique},
   journal={Journal of open source software},
   doi={10.21105/joss.03139},
