### 2025-11-11 310a490
README.md
Updated citation format and details for AssumpSure.
