+++ DESCRIPTION
@@ -6,7 +6,6 @@ Authors@R:
 Description: A 'shiny' application to assess statistical assumptions and guide users toward appropriate tests. The app is designed for researchers with minimal statistical training and provides diagnostics, plots, and test recommendations for a wide range of analyses. Many statistical assumptions are implemented using the package 'rstatix' (Kassambara, 2019) <doi:10.32614/CRAN.package.rstatix> and 'performance' (Lüdecke et al., 2021) <doi:10.21105/joss.03139>.
 License: GPL-3
 Imports:
-Imports:
     shiny,
     shinyjs,
     shinyBS,
