+++ README.md
@@ -3,8 +3,7 @@
 [![](https://www.r-pkg.org/badges/version/AssumpSure?color=orange)](https://cran.r-project.org/package=AssumpSure) ![License: GPL-3](https://img.shields.io/badge/license-GPL--3-blue.svg) [![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active) 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17539689.svg)](https://doi.org/10.5281/zenodo.17539689) 
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.09286/status.svg)](https://doi.org/10.21105/joss.09286)
-![Downloads](https://img.shields.io/github/downloads/Ahmedbargheet/AssumpSure/total?label=Downloads&color=blue)
-
+![CRAN Downloads](https://cranlogs.r-pkg.org/badges/AssumpSure)
 
 
 ## 🎯 Purpose
